Fixes the entry points
